refactor access to shepherd hardware
