+++ software/python-package/shepherd_sheep/h5_recorder_gpio.py
@@ -7,7 +7,7 @@ from shepherd_core.data_models.task import Compression
 from .h5_monitor_abc import Monitor
 from .shared_mem_gpio_output import GPIOTrace
 from .shared_mem_gpio_output import SharedMemGPIOOutput
-from .target_io import GPIO_LOG_BIT_POSITIONS
+from .hardware_target_io import GPIO_LOG_BIT_POSITIONS
 
 
 class GpioRecorder(Monitor):
